+++ fog/report/validation/README.md
@@ -1,6 +1,6 @@
 mc-fog-report-validation
 ========================
 
-This contains the logic for validating a report created by an ingest node
+This contains the logic for validating results from the fog report server
 and extracting the validated pubkey. This crate must not depend on grpcio and
 must be libmobilecoin-friendly.
