+++ fog/report/validation/test-utils/README.md
@@ -1,0 +1,5 @@
+mc-fog-report-validation-test-utils
+===================================
+
+This crate contains a mock fog resolver object for use with the transaction
+builder in unit tests. It bypasses IAS, x509, and grpc, and must not be used in production.
