+++ mobilecoind/Cargo.toml
@@ -62,6 +62,7 @@ tempdir = "0.3"
 mc-common = { path = "../common", features = ["loggers"] }
 mc-connection-test-utils = { path = "../connection/test-utils" }
 mc-fog-report-validation = { path = "../fog/report/validation", features = ["automock"] }
+mc-fog-report-validation-test-utils = { path = "../fog/report/validation/test-utils" }
 mc-transaction-core-test-utils = { path = "../transaction/core/test-utils" }
 mc-util-from-random = { path = "../util/from-random" }
