+++ transaction/core/test-utils/Cargo.toml
@@ -12,7 +12,7 @@ tempdir = "0.3"
 mc-account-keys = { path = "../../../account-keys" }
 mc-crypto-keys = { path = "../../../crypto/keys", default-features = false }
 mc-crypto-rand = { path = "../../../crypto/rand" }
-mc-fog-report-validation = { path = "../../../fog/report/validation" }
+mc-fog-report-validation-test-utils = { path = "../../../fog/report/validation/test-utils" }
 mc-ledger-db = { path = "../../../ledger/db" }
 mc-transaction-core = { path = "../../../transaction/core" }
 mc-transaction-std = { path = "../../../transaction/std" }
