+++ transaction/core/test-utils/src/lib.rs
@@ -4,7 +4,7 @@ use core::convert::TryFrom;
 pub use mc_account_keys::{AccountKey, PublicAddress, ViewKey, DEFAULT_SUBADDRESS_INDEX};
 use mc_crypto_keys::{RistrettoPrivate, RistrettoPublic};
 use mc_crypto_rand::{CryptoRng, RngCore};
-pub use mc_fog_report_validation::MockFogResolver;
+pub use mc_fog_report_validation_test_utils::MockFogResolver;
 use mc_ledger_db::{Ledger, LedgerDB};
 pub use mc_transaction_core::{
     constants::MINIMUM_FEE,
